# schedule-service/docs/swagger.yaml
basePath: /raspyx/api
definitions:
...
  models.AddTeacherRequest:
    properties:
      first_name:
        example: Имя
        type: string
      middle_name:
        example: Отчество
        type: string
      second_name:
        example: Фамилия
        type: string
    type: object
...
  models.ErrorResponse:
    properties:
      code:
        type: integer
      message:
        type: string
    type: object
  models.GetTeacherResponse:
    properties:
      full_name:
        example: ФИО
        type: string
      uuid:
        example: 4b6c34bd-01f0-4fbe-be65-217984b3e33d
        type: string
    type: object
...
  models.ResponseAPI:
    properties:
      errors:
        $ref: '#/definitions/models.ErrorResponse'
      message:
        type: string
      requestId:
        type: string
      result: {}
      success:
        type: boolean
    type: object
...
  models.TeacherResponse:
    properties:
      teacher_uuid:
        example: 4b6c34bd-01f0-4fbe-be65-217984b3e33d
        type: string
    type: object
...
  models.UpdateTeacherRequest:
    properties:
      first_name:
        example: Имя
        type: string
      middle_name:
        example: Отчество
        type: string
      second_name:
        example: Фамилия
        type: string
    type: object
...
host: localhost:8080
info:
  contact: {}
  description: API for schedules
  title: Raspyx
  version: 2.0.0
paths:
...
  /v2/teachers:
    get:
      consumes:
      - application/json
      description: Получение списка всех преподавателей в системе
      produces:
      - application/json
      responses:
        "200":
          description: Данные о преподавателях успешно получены
          schema:
            allOf:
            - $ref: '#/definitions/models.ResponseAPI'
            - properties:
                result:
                  items:
                    $ref: '#/definitions/models.GetTeacherResponse'
                  type: array
              type: object
        "500":
          description: Внутренняя ошибка сервера
          schema:
            $ref: '#/definitions/models.ResponseAPI'
      security:
      - ApiKeyAuth: []
      summary: Получение преподавателей
      tags:
      - teachers
    post:
      consumes:
      - application/json
      description: Создание нового преподавателя
      parameters:
      - description: Данные нового преподавателя
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/models.AddTeacherRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Преподаватель создана
          schema:
            allOf:
            - $ref: '#/definitions/models.ResponseAPI'
            - properties:
                result:
                  $ref: '#/definitions/models.TeacherResponse'
              type: object
        "400":
          description: Неверный формат запроса
          schema:
            $ref: '#/definitions/models.ResponseAPI'
        "409":
          description: Конфликт при создании преподавателя
          schema:
            $ref: '#/definitions/models.ResponseAPI'
        "500":
          description: Внутренняя ошибка сервера
          schema:
            $ref: '#/definitions/models.ResponseAPI'
      security:
      - ApiKeyAuth: []
      summary: Добавление преподавателя
      tags:
      - teachers
  /v2/teachers/{teacher_uuid}:
    delete:
      consumes:
      - application/json
      description: Удаление преподавателя из системы по его идентификатору
      parameters:
      - description: Идентификатор преподавателя
        in: path
        name: teacher_uuid
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Преподаватель успешно удалена
          schema:
            allOf:
            - $ref: '#/definitions/models.ResponseAPI'
            - properties:
                result:
                  $ref: '#/definitions/models.TeacherResponse'
              type: object
        "400":
          description: Неверный формат запроса
          schema:
            $ref: '#/definitions/models.ResponseAPI'
        "404":
          description: Преподаватель не найден
          schema:
            $ref: '#/definitions/models.ResponseAPI'
        "409":
          description: Конфликт при удалении преподавателя
          schema:
            $ref: '#/definitions/models.ResponseAPI'
        "500":
          description: Внутренняя ошибка сервера
          schema:
            $ref: '#/definitions/models.ResponseAPI'
      security:
      - ApiKeyAuth: []
      summary: Удаление преподавателя
      tags:
      - teachers
    put:
      consumes:
      - application/json
      description: Обновление информации о преподавателе по его идентификатору
      parameters:
      - description: Идентификатор преподавателя
        in: path
        name: teacher_uuid
        required: true
        type: string
      - description: Данные для обновления преподавателя
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/models.UpdateTeacherRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Данные о преподавателе успешно обновлены
          schema:
            allOf:
            - $ref: '#/definitions/models.ResponseAPI'
            - properties:
                result:
                  $ref: '#/definitions/models.TeacherResponse'
              type: object
        "400":
          description: Неверный формат запроса
          schema:
            $ref: '#/definitions/models.ResponseAPI'
        "404":
          description: Преподаватель не найден
          schema:
            $ref: '#/definitions/models.ResponseAPI'
        "409":
          description: Конфликт при обновлении преподавателя
          schema:
            $ref: '#/definitions/models.ResponseAPI'
        "500":
          description: Внутренняя ошибка сервера
          schema:
            $ref: '#/definitions/models.ResponseAPI'
      security:
      - ApiKeyAuth: []
      summary: Обновление преподавателя
      tags:
      - teachers
  /v2/teachers/fio/{teacher_fio}:
    get:
      consumes:
      - application/json
      description: Получение информации о преподавателе по его teacher_fio
      parameters:
      - description: ФИО преподавателя
        in: path
        name: teacher_fio
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Данные о преподавателе успешно получены
          schema:
            allOf:
            - $ref: '#/definitions/models.ResponseAPI'
            - properties:
                result:
                  items:
                    $ref: '#/definitions/models.GetTeacherResponse'
                  type: array
              type: object
        "404":
          description: Преподаватель не найден
          schema:
            $ref: '#/definitions/models.ResponseAPI'
        "500":
          description: Внутренняя ошибка сервера
          schema:
            $ref: '#/definitions/models.ResponseAPI'
      security:
      - ApiKeyAuth: []
      summary: Получение преподавателя по его ФИО
      tags:
      - teachers
  /v2/teachers/uuid/{teacher_uuid}:
    get:
      consumes:
      - application/json
      description: Получение информации о преподавателе по её teacher_uuid
      parameters:
      - description: Идентификатор преподавателя
        in: path
        name: teacher_uuid
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Данные о преподавателе успешно получены
          schema:
            allOf:
            - $ref: '#/definitions/models.ResponseAPI'
            - properties:
                result:
                  $ref: '#/definitions/models.GetTeacherResponse'
              type: object
        "404":
          description: Преподаватель не найден
          schema:
            $ref: '#/definitions/models.ResponseAPI'
        "500":
          description: Внутренняя ошибка сервера
          schema:
            $ref: '#/definitions/models.ResponseAPI'
      security:
      - ApiKeyAuth: []
      summary: Получение преподавателя по идентификатору
      tags:
      - teachers
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"